Move default data ID logic to its own file
This is a bit of a chunk and it's unrelated to anything else in _query_plan.py.
